Chat room msg

+++ base/views.py
@@ -6,7 +6,7 @@ from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.forms import UserCreationForm
 from django.http import HttpResponse
-from .models import Room, Topic
+from .models import Room, Topic, Message
 from .forms import RoomForm
 
 #rooms= [
@@ -15,9 +15,7 @@ from .forms import RoomForm
 #    {'id':3, 'name':'Frontend developer'},
 # ]
 
-
 # Create your views here.
-
 def loginPage(request):
     page = 'login' 
     if request.user.is_authenticated:
@@ -84,7 +82,19 @@ def home(request):
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    context = {'room': room}
+    room_messages = room.message_set.all().order_by('-created')
+    participants = room.participants.all()
+    
+    if request.method == "POST":
+        message = Message.objects.create (
+            user=request.user,
+            room=room,
+            body=request.POST.get('body')   
+            )
+        room.participants.all()
+        return redirect('room', pk=room.id)
+    
+    context = {'room': room, 'room_messages': room_messages, 'participants':participants}
     return render(request, 'base/room.html', context)
    
 @login_required(login_url='login')    
@@ -122,7 +132,7 @@ def updateRoom(request, pk):
     form = RoomForm(instance=room)
     topics =Topic.objects.all()
 
-    if request.user != room.host or request.user:
+    if request.user != room.host:
             return HttpResponse('You are not allowed here!!')
 
     if request.method == 'POST' :
@@ -141,8 +151,25 @@ def updateRoom(request, pk):
 @login_required(login_url='login') 
 def deleteRoom(request, pk) :
     room = Room.objects.get(id=pk)
+    
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render (request, 'base/delete.html', {'obj': room})
+
+
+@login_required(login_url='login') 
+def deleteMessage(request, pk) :
+    message = Message.objects.get(id=pk)
+    
+    if request.user != message.user:
+        return HttpResponse('You are not allowed here!!')
+    
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+    return render (request, 'base/delete.html', {'obj': message})
